[DEAF-300] finish export to file dialog refactoring
