+++ app/components/wallet/WalletSettings.js
@@ -11,8 +11,8 @@ import ReadOnlyInput from '../widgets/forms/ReadOnlyInput';
 import DeleteWalletButton from './settings/DeleteWalletButton';
 import DeleteWalletConfirmationDialog from './settings/DeleteWalletConfirmationDialog';
 import DeleteWalletDialogContainer from '../../containers/wallet/dialogs/DeleteWalletDialogContainer';
-import WalletExportDialog from './settings/export/WalletExportToFileDialog';
-import WalletExportDialogContainer from '../../containers/wallet/dialogs/WalletExportToFileDialogContainer';
+import WalletExportDialog from './settings/export-to-file/WalletExportToFileDialog';
+import WalletExportDialogContainer from '../../containers/wallet/settings/WalletExportToFileDialogContainer';
 import ExportPaperWalletPrinterCopyDialog from './settings/paper-wallet-export-dialogs/ExportPaperWalletPrinterCopyDialog';
 import ExportPaperWalletPrinterCopyDialogContainer from '../../containers/wallet/dialogs/paper-wallet-export/ExportPaperWalletPrinterCopyDialogContainer';
 import ExportPaperWalletMnemonicDialog from './settings/paper-wallet-export-dialogs/ExportPaperWalletMnemonicDialog';
